+++ fxmanifest.lua
@@ -3,5 +3,6 @@ resource_type 'gametype' { name = 'Survive the Hunt' }
 
 client_scripts { 'SurviveTheHuntClient.net.dll' }
 server_scripts { 'SurviveTheHuntServer.net.dll' }
+shared_scripts { 'SurviveTheHuntShared.net.dll' }
 
 game 'gta5'