+++ src/SurviveTheHuntShared/Core/Teams.cs
@@ -1,0 +1,7 @@
+﻿namespace SurviveTheHuntShared.Core
+{
+    public static class Teams
+    {
+        public enum Team { Hunters = 0, Hunted };
+    }
+}